+++ R/zzz.R
@@ -6,5 +6,7 @@ globalVariables(c("id", "label_en", "new_id", "has_broader", "external",
                   "has_close_match", "has_exact_match", "has_broader_match",
                   "new_exact_match", "new_narrower_match", "new_close_match",
                   "new_broader_match", "avail", "has_source", "label", "newid",
-                  "external_label"
+                  "external_label", "harmLab", "sort_in", "verbose", "new_label",
+                  "prev_broader_match", "prev_close_match", "prev_exact_match",
+                  "prev_narrower_match", "children"
                   ))